requirements.txt
antlr4-python3-runtime==4.9.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
contourpy==1.3.2
cubinlinker-cu11==0.3.0.post2
cuda-bindings==11.8.6
cuda-python==11.8.6
cudf-cu11==25.4.0
cuml-cu11==25.4.0
cupy-cuda11x==13.4.1
cuvs-cu11==25.4.0
cycler==0.12.1
dask==2025.2.0
dask-cuda==25.4.0
dask-cudf-cu11==25.4.0
distributed==2025.2.0
distributed-ucxx-cu11==0.43.0
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
idna==3.10
importlib_metadata==8.6.1
iopath==0.1.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
libcudf-cu11==25.4.0
libcuml-cu11==25.4.0
libcuvs-cu11==25.4.0
libkvikio-cu11==25.4.0
libraft-cu11==25.4.0
librmm-cu11==25.4.0
libucx-cu11==1.18.0.post1
libucxx-cu11==0.43.0
lit==18.1.8
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.60.0
numba-cuda==0.4.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.570.86
nvidia-nvcomp-cu11==4.2.0.11
nvtx==0.2.11
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
partd==1.4.2
pillow==11.2.1
portalocker==3.1.1
psutil==7.0.0
ptxcompiler-cu11==0.8.1.post2
pyarrow==19.0.1
Pygments==2.19.1
pylibcudf-cu11==25.4.0
pylibraft-cu11==25.4.0
pynvml==12.0.0
pyparsing==3.2.3
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
raft-dask-cu11==25.4.0
rapids-dask-dependency==25.4.0
rapids-logger==0.1.1
requests==2.32.3
rich==14.0.0
rmm-cu11==25.4.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sortedcontainers==2.4.0
submitit==1.5.2
sympy==1.13.3
tabulate==0.9.0
tblib==3.1.0
termcolor==3.0.1
threadpoolctl==3.6.0
toolz==1.0.0
torch==2.0.0+cu117
torchmetrics==0.10.3
torchvision==0.15.0+cu117
tornado==6.4.2
tqdm==4.67.1
treelite==4.4.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
ucx-py-cu11==0.43.0
ucxx-cu11==0.43.0
urllib3==2.4.0
xformers==0.0.18
yacs==0.1.8
zict==3.0.0
zipp==3.21.0
